requirements.txt
adafruit-circuitpython-neopixel
rpi-ws281x==5.0.0
